requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@f24f34e3f1fc350664cc93b1f57510bd9c714a7b
DiscordUtils==1.3.4
frozenlist==1.3.0
humanize==4.1.0
idna==3.3
multidict==6.0.2
Pillow==9.1.1
pycodestyle==2.8.0
pytimeparse==1.1.8
toml==0.10.2
typing_extensions==4.2.0
yarl==1.7.2
